+++ src/Language/Futhark/Interpreter.hs
@@ -30,7 +30,7 @@ import Control.Monad.Reader
 import Control.Monad.State
 import Control.Monad.Trans.Maybe
 import Data.Array
-import Data.Bifunctor (first)
+import Data.Bifunctor (first, second)
 import Data.List
   ( find,
     foldl',
@@ -1110,19 +1110,27 @@ evalCase v env (CasePat p cExp _) = runMaybeT $ do
   env' <- patternMatch env p v
   lift $ eval env' cExp
 
+-- We hackily do multiple substitutions in modules, because otherwise
+-- we would lose in cases where the parameter substitutions are [a->x,
+-- b->x] when we reverse. (See issue #1250.)
+reverseSubstitutions :: M.Map VName VName -> M.Map VName [VName]
+reverseSubstitutions =
+  M.fromListWith (<>) . map (second pure . uncurry (flip (,))) . M.toList
+
 substituteInModule :: M.Map VName VName -> Module -> Module
 substituteInModule substs = onModule
   where
     rev_substs = reverseSubstitutions substs
-    replace v = fromMaybe v $ M.lookup v rev_substs
-    replaceQ v = maybe v qualName $ M.lookup (qualLeaf v) rev_substs
+    replace v = fromMaybe [v] $ M.lookup v rev_substs
+    replaceQ v = maybe v qualName $ maybeHead =<< M.lookup (qualLeaf v) rev_substs
     replaceM f m = M.fromList $ do
       (k, v) <- M.toList m
-      return (replace k, f v)
+      k' <- replace k
+      return (k', f v)
     onModule (Module (Env terms types _)) =
       Module $ Env (replaceM onTerm terms) (replaceM onType types) mempty
     onModule (ModuleFun f) =
-      ModuleFun $ \m -> onModule <$> f (substituteInModule rev_substs m)
+      ModuleFun $ \m -> onModule <$> f (substituteInModule (M.mapMaybe maybeHead rev_substs) m)
     onTerm (TermValue t v) = TermValue t v
     onTerm (TermPoly t v) = TermPoly t v
     onTerm (TermModule m) = TermModule $ onModule m
@@ -1130,9 +1138,6 @@ substituteInModule substs = onModule
     onDim (NamedDim v) = NamedDim $ replaceQ v
     onDim (ConstDim x) = ConstDim x
     onDim AnyDim = AnyDim
-
-reverseSubstitutions :: M.Map VName VName -> M.Map VName VName
-reverseSubstitutions = M.fromList . map (uncurry $ flip (,)) . M.toList
 
 evalModuleVar :: Env -> QualName VName -> EvalM Module
 evalModuleVar env qv =